CDS: Support config `agent.span_limit_per_segment`
